employee can attempt test

+++ app/Http/Controllers/QuestionController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\Models\Course;
-use App\Models\Option;
 use App\Models\Question;
 use App\Models\Test;
 use App\Models\Unit;
@@ -27,9 +26,14 @@ class QuestionController extends Controller
     {
         $attributes = $request->validate([
             'question' => 'required|min:3',
-            'options' => 'array|size:2',
+            'options' => 'array|min:2',
             'options.*' => 'required|min:3|max:255',
             'radio' => 'required|min:1|gt:0'
+        ], [
+            'options.*.required' => 'The Option field is required',
+            'options.*.min' => 'The minimum lenght should be of 3 characters',
+            'options.*.max' => 'The maximum lenght should be of 255 characters',
+            'radio.required' => 'Answer field is required'
         ]);
 
         $question = Question::create([
@@ -38,29 +42,22 @@ class QuestionController extends Controller
 
         $question->test()->attach($test);
 
-       $options = collect($attributes['options']);
+        $i =1;
+        collect($attributes['options'])
+            ->each(function($option) use($question, &$i, $attributes) {
 
-       $i =0;
-
-       $options->each(function($option) use($question, &$i, $attributes) {
-            $i++;
-
-            if ($i == $attributes['radio'])
-            {
-                Option::create([
-                    'question_id' => $question->id,
-                    'option' => $option,
-                    'answer' => true
-                ]);
-            }
-            else
-            {
-                Option::create([
-                    'question_id' => $question->id,
+                $answer = $question->options()->create([
                     'option' => $option
                 ]);
-            }
-        });
+
+                if ($i == $attributes['radio'])
+                {
+                   $answer->update([
+                        'answer' => true
+                    ]);
+                }
+                $i++;
+            });
 
         if($request['save'] == 'save')
         {
@@ -72,6 +69,8 @@ class QuestionController extends Controller
 
     public function edit(Course $course, Unit $unit, Test $test, Question $question)
     {
+        $this->authorize('update', $course);
+
         return view('trainers.courses.units.test.question.edit', [
             'course' => $course,
             'unit' => $unit,
@@ -80,13 +79,20 @@ class QuestionController extends Controller
         ]);
     }
 
-    public function update(Request $request, Question $question)
+    public function update(Request $request, Course $course, Question $question)
     {
+        $this->authorize('update', $course);
+        
         $request->validate([
             'question' => 'required|min:3',
             'options' => 'array|size:2',
             'options.*' => 'required|min:3|max:255',
             'radio' => 'required|min:1|gt:0'
+        ],  [
+            'options.*.required' => 'The Option field is required',
+            'options.*.min' => 'The minimum lenght should be of 3 characters',
+            'options.*.max' => 'The maximum lenght should be of 255 characters',
+            'radio.required' => 'Answer field is required'
         ]);
 
         $question->update([
@@ -118,8 +124,10 @@ class QuestionController extends Controller
 
     }
 
-    public function destroy(Question $question)
+    public function destroy(Course $course ,Question $question)
     {
+        $this->authorize('update', $course);
+        
         $question->delete();
 
         return back()->with('status', 'Successfully Deleted');
