+++ app/Http/Controllers/TestController.php
@@ -46,6 +46,8 @@ class TestController extends Controller
 
     public function edit(Course $course, Unit $unit, Test $test)
     {
+        $this->authorize('update', $course);
+
         return view('trainers.courses.units.test.edit', [
             'course' => $course,
             'unit' => $unit,
@@ -53,21 +55,38 @@ class TestController extends Controller
         ]);
     }
 
-    public function update(Request $request, Test $test)
+    public function update(Request $request, Course $course, Unit $unit,Test $test)
     {
+        $this->authorize('update', $course);
+
         $attributes = $request->validate([
             'name' => 'required|min:3|max:255',
             'duration' => 'required|numeric|gt:0',
             'pass_score' => 'required|numeric|between:1,100'
         ]);
+        if ($test->duration < $attributes['duration'])
+        {
+            $unit->decrement('duration', $test->duration);
+            $unit->increment('duration', $attributes['duration']);
+        }
+        elseif ($test->duration > $attributes['duration'])
+        {
+            $unit->decrement('duration', $test->duration);
+            $unit->increment('duration', $attributes['duration']);
+        }
 
         $test->update($attributes);
+        $test->lessons()->update([
+            'duration' => $attributes['duration']
+        ]);
 
         return back()->with('status', 'Successfully Test Updated');
     }
 
-    public function destroy(Test $test)
+    public function destroy(Course $course, Test $test)
     {
+        $this->authorize('delete', $course);
+
         $test->delete($test);
 
         return back()->with('status', 'Successfully Test Updated');
